requirements.txt
asgiref==3.7.2
certifi==2024.8.30
charset-normalizer==3.3.2
Django==4.2.4
djangorestframework==3.15.2
idna==3.10
inflection==0.5.1
Pillow==10.0.1
psycopg2==2.9.9
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.2.3
